start on a test for blocking peers.
I had to write way to much code and do too much debugging
to get to this point. That's a problem for potential developers,
this should be simple.

One problem was that connect is duplicated between client
and sbot#connect, and I wasn't sure which was being used.
Also, I'm realizing that sbot could be more generic. The
specific way that follows work is really a phoenix/patchwork
feature, sbot should really just provide the hooks that
make implementing that easy. This is going to become more
apparent as blocking is developed, I think.