Add tests for bin.js (sbot binary)
- test for config via command line (host and port, ipv4 + ipv6)
- test for config via config file
- test for bin.js with ipv6 host in local config
- sbot should provide a multiserver address containing ws:

fixes #552 and closes #556

notes from squashed commits:

> This prevents ssb-config from reading ~/.ssb/config

Travis CI does not support IPv6 at the moment.
Source: https://blog.travis-ci.com/2015-11-27-moving-to-a-more-elastic-future

> test/bin.js: add dialing ws endpoint and call something

@regular argues this is too much integration testing and I partially agree
but cryptix argues to keep this in until ssb-ws has tests of it's own.

(https://github.com/ssbc/scuttlebot/pull/557)

> Using bash job control to manage sbot lifetime in tests

Add a helper funtion that runs sbot in a subshell and returns a
function to kill that sbot instance and then end the test.

This simplifies testing bin.js because now each test cleans up after
iteself and it is guaranteed that only one sbot is running at a time.

Add {igore:} option to try_often. If set, errors that don't match the
specified regex will make the test fail immediately.

> Fix ws test, use multiserver-address, random shs.caps