Add missing dependencies, dynamically for child deps
This commit makes two changes: the least controversial change is to add
two modules to `package.json` which are required in our test files.

The second change is to add the necessary devDependencies for child
dependencies when we run their tests. Specifically, this means that
we're installing all `devDependencies` except those that have been
included as production dependencies in the ssb-server package.

This was originally added in #617, but when #618 was merged it hardcoded
these dependencies instead of adding them dynamically. A few commits
later these hardcoded child devdeps were removed in 95785cb, which means
that `npm test` started failing again. This commit resolves the problem.