Replace `getAddress()` with `address()`
Problem: The `getAddress()` MuxRPC calls seem to be failing, but it's
unclear when the problem started or why. The Secret-Stack documentation
specifies a `getAddress()` method, but it seems to only be available via
an in-process Node.js method call rather than a MuxRPC method call. For
some reason, the MuxRPC method call (which implements the same behavior)
is called `address` rather than `getAddress`.

Solution: Replace all instances of the `getAddress()` MuxRPC method call
with the exact same method call to `address()`. This *is* a breaking
change, but it's a breaking change that already existed in previous
versions of SSB-Server (!) and the next version of SSB-Server needs to
have a major version bump anyway since FlumeDB and SSB-DB have had
breaking changes. This commit also reverts recent changes to
`test/bin.js`, which were causing test failures.